updated paths and scripts
