+++ index.html
@@ -86,7 +86,17 @@
         <span class="">Gettr</span>
       </div>
     </footer>
-    <script src="node_modules/bootstrap/dist/js/bootstrap.js"></script>
+    <script
+      src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.11.6/dist/umd/popper.min.js"
+      integrity="sha384-oBqDVmMz9ATKxIep9tiCxS/Z9fNfEXiDAYTujMAeBAsjFuCZSmKbSSUnQlmh/jp3"
+      crossorigin="anonymous"
+    ></script>
+    <script
+      src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/js/bootstrap.min.js"
+      integrity="sha384-IDwe1+LCz02ROU9k972gdyvl+AESN10+x7tBKgc9I5HFtuNz0wWnPclzo6p9vxnk"
+      crossorigin="anonymous"
+    ></script>
+    <script src="/node_modules/bootstrap/dist/js/bootstrap.js"></script>
     <script type="module" src="/javascript/api/actions/index.js"></script>
     <script type="module" src="/javascript/api/actions/functions/createPost.js"></script>
     <script src="/javascript/api/actions/functions/filterpost.js"></script>
