+++ javascript/api/actions/functions/filterpost.js
@@ -1,6 +1,10 @@
 let params = "limit=10&_author=true";
 let url = `https://nf-api.onrender.com/api/v1/social/posts?${params}`;
 
+/**
+ * @param {object} getPosts - The function that gets the posts.
+ */
+
 const getPosts = async () => {
   const token = localStorage.getItem("token") || sessionStorage.getItem("token");
   const response = await fetch(url, {
@@ -40,6 +44,10 @@ const getPosts = async () => {
     });
   });
 };
+
+/**
+ * @param {string} filter // filter posts by the options "newest", "oldest", "A-Z", "Z-A"
+ */
 
 function filterPosts(filter) {
   params = "limit=10&_author=true";
